removed old gen_bipartite
